Added file compare - reset when no progress
Ran the numbers running for a full hour 3 times, and the local maximum appears to be at the major reset when progress then stops until the minors catch up. Changing so ascends happen when progress on APs has halted for 3 minutes. Also shrank file size to just the AP so I can do OCR next instead of file compare.